replaced all int/size_t indices in wxControlWithItems API with unsigned int (committing on behalf of ABX)
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@38319 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ include/wx/checklst.h
@@ -32,8 +32,8 @@ public:
     wxCheckListBoxBase() { }
 
     // check list box specific methods
-    virtual bool IsChecked(size_t item) const = 0;
-    virtual void Check(size_t item, bool check = true) = 0;
+    virtual bool IsChecked(unsigned int item) const = 0;
+    virtual void Check(unsigned int item, bool check = true) = 0;
 
     DECLARE_NO_COPY_CLASS(wxCheckListBoxBase)
 };
