+++ include/wx/mac/carbon/listbox.h
@@ -75,11 +75,11 @@ public:
 
     // implement base class pure virtuals
     virtual void Clear();
-    virtual void Delete(int n);
+    virtual void Delete(unsigned int n);
 
-    virtual size_t GetCount() const;
-    virtual wxString GetString(int n) const;
-    virtual void SetString(int n, const wxString& s);
+    virtual unsigned int GetCount() const;
+    virtual wxString GetString(unsigned int n) const;
+    virtual void SetString(unsigned int n, const wxString& s);
     virtual int FindString(const wxString& s, bool bCase = false) const;
 
     virtual bool IsSelected(int n) const;
@@ -105,13 +105,13 @@ public:
 protected:
     virtual void DoSetSelection(int n, bool select);
     virtual int DoAppend(const wxString& item);
-    virtual void DoInsertItems(const wxArrayString& items, int pos);
+    virtual void DoInsertItems(const wxArrayString& items, unsigned int pos);
     virtual void DoSetItems(const wxArrayString& items, void **clientData);
     virtual void DoSetFirstItem(int n);
-    virtual void DoSetItemClientData(int n, void* clientData);
-    virtual void* DoGetItemClientData(int n) const;
-    virtual void DoSetItemClientObject(int n, wxClientData* clientData);
-    virtual wxClientData* DoGetItemClientObject(int n) const;
+    virtual void DoSetItemClientData(unsigned int n, void* clientData);
+    virtual void* DoGetItemClientData(unsigned int n) const;
+    virtual void DoSetItemClientObject(unsigned int n, wxClientData* clientData);
+    virtual wxClientData* DoGetItemClientObject(unsigned int n) const;
     virtual void DoSetSize(int x, int y, int width, int height, int sizeFlags = wxSIZE_AUTO);
     virtual int DoListHitTest(const wxPoint& point) const;
 
@@ -132,7 +132,7 @@ protected:
     // prevent collision with some BSD definitions of macro Free()
     void FreeData();
 
-    size_t m_noItems;
+    unsigned int m_noItems;
     int m_selected;
     bool m_suppressSelection ;
     wxString  m_typeIn ;
