+++ include/wx/mac/classic/combobox.h
@@ -78,15 +78,15 @@ class WXDLLEXPORT wxComboBox : public wxControl, public wxComboBoxBase
            const wxString& name = wxComboBoxNameStr);
 
     // List functions
-    virtual void Delete(int n);
+    virtual void Delete(unsigned int n);
     virtual void Clear();
 
     virtual int GetSelection() const ;
     virtual void SetSelection(int n);
     virtual int FindString(const wxString& s, bool bCase = false) const;
-    virtual wxString GetString(int n) const ;
+    virtual wxString GetString(unsigned int n) const ;
     virtual wxString GetStringSelection() const ;
-    virtual void SetString(int n, const wxString& s) ;
+    virtual void SetString(unsigned int n, const wxString& s);
 
     // Text field functions
     virtual wxString GetValue() const ;
@@ -104,7 +104,7 @@ class WXDLLEXPORT wxComboBox : public wxControl, public wxComboBoxBase
     virtual void Remove(long from, long to);
     virtual void SetSelection(long from, long to);
     virtual void SetEditable(bool editable);
-    virtual size_t GetCount() const { return m_choice->GetCount() ; }
+    virtual unsigned int GetCount() const { return m_choice->GetCount() ; }
 
     virtual bool IsEditable() const ;
 
@@ -124,12 +124,12 @@ class WXDLLEXPORT wxComboBox : public wxControl, public wxComboBoxBase
 
 protected:
     virtual int DoAppend(const wxString& item) ;
-    virtual int DoInsert(const wxString& item, int pos) ;
+    virtual int DoInsert(const wxString& item, unsigned int pos) ;
 
-    virtual void DoSetItemClientData(int n, void* clientData) ;
-    virtual void* DoGetItemClientData(int n) const ;
-    virtual void DoSetItemClientObject(int n, wxClientData* clientData) ;
-    virtual wxClientData* DoGetItemClientObject(int n) const ;
+    virtual void DoSetItemClientData(unsigned int n, void* clientData);
+    virtual void* DoGetItemClientData(unsigned int n) const ;
+    virtual void DoSetItemClientObject(unsigned int n, wxClientData* clientData);
+    virtual wxClientData* DoGetItemClientObject(unsigned int n) const ;
 
     void FreeData();
