+++ include/wx/msw/radiobox.h
@@ -85,13 +85,13 @@ public:
     // implement the radiobox interface
     virtual void SetSelection(int n);
     virtual int GetSelection() const { return m_selectedButton; }
-    virtual size_t GetCount() const;
-    virtual wxString GetString(int n) const;
-    virtual void SetString(int n, const wxString& label);
-    virtual bool Enable(int n, bool enable = true);
-    virtual bool Show(int n, bool show = true);
-    virtual bool IsItemEnabled(int n) const;
-    virtual bool IsItemShown(int n) const;
+    virtual unsigned int GetCount() const;
+    virtual wxString GetString(unsigned int n) const;
+    virtual void SetString(unsigned int n, const wxString& label);
+    virtual bool Enable(unsigned int n, bool enable = true);
+    virtual bool Show(unsigned int n, bool show = true);
+    virtual bool IsItemEnabled(unsigned int n) const;
+    virtual bool IsItemShown(unsigned int n) const;
 
     // override some base class methods
     virtual bool Show(bool show = true);
