+++ include/wx/os2/checklst.h
@@ -49,17 +49,15 @@ public:
     //
     // Override base class virtuals
     //
-    virtual void Delete(int n);
+    virtual void Delete(unsigned int n);
 
     virtual bool SetFont(const wxFont &rFont);
 
     //
     // Items may be checked
     //
-    bool IsChecked(size_t uiIndex) const;
-    void Check( size_t uiIndex
-               ,bool   bCheck = true
-              );
+    bool IsChecked(unsigned int uiIndex) const;
+    void Check(unsigned int uiIndex, bool bCheck = true);
 
     //
     // Accessors
@@ -74,7 +72,7 @@ protected:
     virtual wxOwnerDrawn* CreateItem(size_t n);
     virtual long          OS2OnMeasure(WXMEASUREITEMSTRUCT* pItem);
 
-    virtual void DoInsertItems(const wxArrayString& items, int pos);
+    virtual void DoInsertItems(const wxArrayString& items, unsigned int pos);
 
     //
     // Pressing space or clicking the check box toggles the item
@@ -83,7 +81,7 @@ protected:
     void OnLeftClick(wxMouseEvent& rEvent);
 
 private:
-    size_t                          m_nItemHeight;  // height of checklistbox items (the same for all)
+    size_t m_nItemHeight;  // height of checklistbox items (the same for all)
 
     DECLARE_DYNAMIC_CLASS(wxCheckListBox)
     DECLARE_EVENT_TABLE()
