+++ include/wx/os2/choice.h
@@ -90,18 +90,16 @@ public:
     // Implement base class virtuals
     //
     virtual int      DoAppend(const wxString& rsItem);
-    virtual int      DoInsert(const wxString& rsItem, int pos);
-    virtual void     Delete(int n);
+    virtual int      DoInsert(const wxString& rsItem, unsigned int pos);
+    virtual void     Delete(unsigned int n);
     virtual void     Clear(void);
 
-    virtual size_t   GetCount() const;
-    virtual int      GetSelection(void) const ;
+    virtual unsigned int GetCount() const;
+    virtual int      GetSelection(void) const;
     virtual void     SetSelection(int n);
 
-    virtual wxString GetString(int n) const ;
-    virtual void     SetString( int n
-                               ,const wxString& rsStr
-                              );
+    virtual wxString GetString(unsigned int n) const;
+    virtual void     SetString(unsigned int n, const wxString& rsStr);
 
     //
     // OS2 only
@@ -115,14 +113,10 @@ public:
                                   );
 
 protected:
-    virtual void          DoSetItemClientData( int   n
-                                              ,void* pClientData
-                                             );
-    virtual void*         DoGetItemClientData(int n) const;
-    virtual void          DoSetItemClientObject( int           n
-                                                ,wxClientData* pClientData
-                                               );
-    virtual wxClientData* DoGetItemClientObject(int n) const;
+    virtual void DoSetItemClientData(unsigned int n, void* pClientData);
+    virtual void*         DoGetItemClientData(unsigned int n) const;
+    virtual void          DoSetItemClientObject(unsigned int n, wxClientData* pClientData);
+    virtual wxClientData* DoGetItemClientObject(unsigned int n) const;
     virtual wxSize        DoGetBestSize(void) const;
     virtual void          DoSetSize( int nX
                                     ,int nY
