+++ include/wx/palmos/radiobox.h
@@ -88,15 +88,15 @@ public:
     // implement the radiobox interface
     virtual void SetSelection(int n);
     virtual int GetSelection() const;
-    virtual size_t GetCount() const;
-    virtual wxString GetString(int n) const;
-    virtual void SetString(int n, const wxString& label);
+    virtual unsigned int GetCount() const;
+    virtual wxString GetString(unsigned int n) const;
+    virtual void SetString(unsigned int n, const wxString& label);
 
     virtual bool Enable(bool enable = true);
-    virtual bool Enable(int n, bool enable = true);
+    virtual bool Enable(unsigned int n, bool enable = true);
 
     virtual bool Show(bool show = true);
-    virtual bool Show(int n, bool show = true);
+    virtual bool Show(unsigned int n, bool show = true);
 
     virtual void SetLabel(const wxString& label);
     virtual wxString GetLabel();
